Promote IO support queries to cudf API
