+++ cpp/include/cudf/io/config_utils.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021-2023, NVIDIA CORPORATION.
+ * Copyright (c) 2024, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,37 +15,10 @@
  */
 #pragma once
 
-#include <cudf/detail/utilities/logger.hpp>
+#include <cudf/utilities/export.hpp>
 
-#include <sstream>
-#include <string>
-
-namespace cudf::io::detail {
-
-/**
- * @brief Returns the value of the environment variable, or a default value if the variable is not
- * present.
- */
-template <typename T>
-T getenv_or(std::string_view env_var_name, T default_val)
-{
-  auto const env_val = std::getenv(env_var_name.data());
-  if (env_val != nullptr) {
-    CUDF_LOG_INFO("Environment variable {} read as {}", env_var_name, env_val);
-  } else {
-    CUDF_LOG_INFO(
-      "Environment variable {} is not set, using default value {}", env_var_name, default_val);
-  }
-
-  if (env_val == nullptr) { return default_val; }
-
-  std::stringstream sstream(env_val);
-  T converted_val;
-  sstream >> converted_val;
-  return converted_val;
-}
-
-namespace cufile_integration {
+namespace CUDF_EXPORT cudf {
+namespace io::cufile_integration {
 
 /**
  * @brief Returns true if cuFile and its compatibility mode are enabled.
@@ -62,9 +35,9 @@ bool is_gds_enabled();
  */
 bool is_kvikio_enabled();
 
-}  // namespace cufile_integration
+}  // namespace io::cufile_integration
 
-namespace nvcomp_integration {
+namespace io::nvcomp_integration {
 
 /**
  * @brief Returns true if all nvCOMP uses are enabled.
@@ -76,6 +49,5 @@ bool is_all_enabled();
  */
 bool is_stable_enabled();
 
-}  // namespace nvcomp_integration
-
-}  // namespace cudf::io::detail
+}  // namespace io::nvcomp_integration
+}  // namespace CUDF_EXPORT cudf
